+++ src/main/java/aDarbellay/s05/t1/model/player/RandomPlayer.java
@@ -24,12 +24,12 @@ public class RandomPlayer implements Player {
     }
 
     @Override
-    public int getId() {
+    public long getId() {
         return 0;
     }
 
     @Override
-    public void setId(int i) {
+    public void setId(long i) {
 
     }
